OA-30: Addressing requested changes
