+++ omod/src/test/java/org/openmrs/module/oauth2login/web/controller/KeycloakAuthenticationTest.java
@@ -34,6 +34,7 @@ public class KeycloakAuthenticationTest extends OAuth2IntegrationTest {
 		Assert.assertEquals("Tommy", user.getGivenName());
 		Assert.assertEquals("Atkins", user.getFamilyName());
 		Assert.assertEquals("tatkins@example.com", user.getEmail());
+		assertProviderAccountActivation(user);
 	}
 	
 	@Override
